Fixed the bug in rptr and empty generation logic - the comparison of the ptrs need to be done on the next_state input to the FSM to get the empty status immediately. Symmetric fix done for wptr and empty logic as well.
A reference RTL working.afifo_rptr* is added which follows the best
practice of separating NS and PS logic - This is also more legible.
For now, I have updated the code within the same process itself, to
find the my own unique solution for the issue. Going forward will
plan to use the structure of the reference RTL itself.